pyproject.toml
[project]
name = "word-mention-func"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flashtext>=2.7",
    "multipledispatch>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "pymystem3>=0.2.0",
    "python-dotenv>=1.1.0",
    "s3p-sdk>=0.2.6",
    "tqdm>=4.67.1",
]
